+++ Puppet/Daemon.hs
@@ -112,7 +112,18 @@ gCatalog :: Preferences IO
 gCatalog prefs getStatements getTemplate stats hquery ndename facts = do
     logDebug ("Received query for node " <> ndename)
     traceEventIO ("START gCatalog " <> T.unpack ndename)
-    (stmts :!: warnings) <- measure stats ndename $ getCatalog interpretMonad getStatements getTemplate (prefs ^. prefPDB) ndename facts (prefs ^. natTypes) (prefs ^. prefExtFuncs) hquery defaultImpureMethods
+    (stmts :!: warnings) <- measure stats ndename $
+                getCatalog interpretMonad
+                           getStatements
+                           getTemplate
+                           (prefs ^. prefPDB)
+                           ndename
+                           facts
+                           (prefs ^. natTypes)
+                           (prefs ^. prefExtFuncs)
+                           hquery
+                           defaultImpureMethods
+                           (prefs ^. ignoredmodules)
     mapM_ (\(p :!: m) -> LOG.logM loggerName p (displayS (renderCompact (ttext ndename <> ":" <+> m)) "")) warnings
     traceEventIO ("STOP gCatalog " <> T.unpack ndename)
     return stmts